Fix FindBugs issue "String + in a loop"
Bug: ContactManagerImpl.toString() concatenates strings using + in a
loop
The method seems to be building a String using concatenation in a loop.
In each iteration, the String is converted to a
StringBuffer/StringBuilder, appended to, and converted back to a String.
This can lead to a cost quadratic in the number of iterations, as the
growing string is recopied in each iteration.

Better performance can be obtained by using a StringBuffer (or
StringBuilder in Java 1.5) explicitly.